refactor(tokens): applying changes to tokens creation services
